Refactor `DeliberationProcessor`
