+++ package.json
@@ -9,10 +9,10 @@
     "gov-app:build": "nx run gov-app:build",
     "gov-app:dev": "nx run gov-app:serve",
     "proposal-relayer:build": "nx run proposal-relayer:build",
+    "proposal-relayer:deliberation-processor": "nx run proposal-relayer:deliberation-processor",
     "proposal-relayer:proposal-listener": "nx run proposal-relayer:proposal-listener",
     "proposal-relayer:proposal-processor": "nx run proposal-relayer:proposal-processor",
-    "proposal-relayer:vote-processor": "nx run proposal-relayer:vote-processor",
-    "proposal-relayer:dev": "run-p proposal-relayer:proposal-listener proposal-relayer:proposal-processor proposal-relayer:vote-processor",
+    "proposal-relayer:dev": "run-p proposal-relayer:proposal-listener proposal-relayer:proposal-processor proposal-relayer:deliberation-processor",
     "new:provider": "HYGEN_TMPLS=.hygen/templates hygen new lib:provider",
     "new:component": "HYGEN_TMPLS=.hygen/templates hygen new lib:component",
     "new:hook": "HYGEN_TMPLS=.hygen/templates hygen new lib:hook",
