+++ packages/proposal-relayer/project.json
@@ -31,11 +31,11 @@
         ]
       }
     },
-    "vote-processor": {
+    "deliberation-processor": {
       "executor": "nx:run-commands",
       "options": {
         "commands": [
-          "ts-node --project packages/proposal-relayer/tsconfig.json -r tsconfig-paths/register packages/proposal-relayer/commands/vote-processor.ts"
+          "ts-node --project packages/proposal-relayer/tsconfig.json -r tsconfig-paths/register packages/proposal-relayer/commands/deliberation-processor.ts"
         ]
       }
     },
